+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Compare with our blacklist too
 - Connect to CSE tools, after compared the list, we well handle cleaned scan and download the result for you.
 - Connect to DataValidation API, right now we can also do DV in this tool right away. we will handle it for you.
+- Able to connect accounts not in the Hi-iQ DB with `promot-customer-selector-manually.js` file.
 ### Changed
 - Remove hard limitation for Email header name, now we can select which hader represent Email column.
+- Change export data to email hash only, fixed the email order position issue.
 ### Fixed
 - Downgrade dependency `async` to version 2.6.3 to fix `doUntil` bug.